+++ addOns/pscanrules/src/main/java/org/zaproxy/zap/extension/pscanrules/ContentSecurityPolicyScanRule.java
@@ -121,10 +121,11 @@ public class ContentSecurityPolicyScanRule extends PluginPassiveScanner {
         if (!xcspOptions.isEmpty()) {
             raiseAlert(
                     Constant.messages.getString(MESSAGE_PREFIX + "xcsp.name"),
-                    Constant.messages.getString(MESSAGE_PREFIX + "xcsp.desc"),
+                    Constant.messages.getString(MESSAGE_PREFIX + "xcsp.otherinfo"),
                     getHeaderField(msg, HTTP_HEADER_XCSP).get(0),
                     cspHeaderFound ? Alert.RISK_INFO : Alert.RISK_LOW,
-                    xcspOptions.get(0));
+                    xcspOptions.get(0),
+                    "1");
         }
 
         // X-WebKit-CSP is supported by Chrome 14+, and Safari 6+
@@ -133,10 +134,11 @@ public class ContentSecurityPolicyScanRule extends PluginPassiveScanner {
         if (!xwkcspOptions.isEmpty()) {
             raiseAlert(
                     Constant.messages.getString(MESSAGE_PREFIX + "xwkcsp.name"),
-                    Constant.messages.getString(MESSAGE_PREFIX + "xwkcsp.desc"),
+                    Constant.messages.getString(MESSAGE_PREFIX + "xwkcsp.otherinfo"),
                     getHeaderField(msg, HTTP_HEADER_WEBKIT_CSP).get(0),
                     cspHeaderFound ? Alert.RISK_INFO : Alert.RISK_LOW,
-                    xwkcspOptions.get(0));
+                    xwkcspOptions.get(0),
+                    "2");
         }
 
         if (cspHeaderFound) {
@@ -170,7 +172,8 @@ public class ContentSecurityPolicyScanRule extends PluginPassiveScanner {
                             cspNoticesString,
                             "",
                             noticesRisk,
-                            csp);
+                            csp,
+                            "3");
                 }
 
                 List<String> allowedWildcardSources = getAllowedWildcardSources(csp);
@@ -181,21 +184,22 @@ public class ContentSecurityPolicyScanRule extends PluginPassiveScanner {
                                     .filter(DIRECTIVES_WITHOUT_FALLBACK::contains)
                                     .collect(Collectors.toList());
                     String allowedWildcardSrcs = String.join(", ", allowedWildcardSources);
-                    String wildcardSrcDesc =
+                    String wildcardSrcOtherInfo =
                             Constant.messages.getString(
-                                    MESSAGE_PREFIX + "wildcard.desc", allowedWildcardSrcs);
+                                    MESSAGE_PREFIX + "wildcard.otherinfo", allowedWildcardSrcs);
                     if (!allowedDirectivesWithoutFallback.isEmpty()) {
-                        wildcardSrcDesc +=
+                        wildcardSrcOtherInfo +=
                                 Constant.messages.getString(
-                                        "pscanrules.csp.desc.extended",
+                                        "pscanrules.csp.otherinfo.extended",
                                         String.join(", ", allowedDirectivesWithoutFallback));
                     }
                     raiseAlert(
                             Constant.messages.getString(MESSAGE_PREFIX + "wildcard.name"),
-                            wildcardSrcDesc,
+                            wildcardSrcOtherInfo,
                             "",
                             Alert.RISK_MEDIUM,
-                            csp);
+                            csp,
+                            "4");
                 }
 
                 Optional<SourceExpressionDirective> optDirective =
@@ -208,10 +212,11 @@ public class ContentSecurityPolicyScanRule extends PluginPassiveScanner {
                                 Constant.messages.getString(
                                         MESSAGE_PREFIX + "scriptsrc.unsafe.name"),
                                 Constant.messages.getString(
-                                        MESSAGE_PREFIX + "scriptsrc.unsafe.desc"),
+                                        MESSAGE_PREFIX + "scriptsrc.unsafe.otherinfo"),
                                 "",
                                 Alert.RISK_MEDIUM,
-                                csp);
+                                csp,
+                                "5");
                     }
                 }
 
@@ -224,10 +229,11 @@ public class ContentSecurityPolicyScanRule extends PluginPassiveScanner {
                                 Constant.messages.getString(
                                         MESSAGE_PREFIX + "stylesrc.unsafe.name"),
                                 Constant.messages.getString(
-                                        MESSAGE_PREFIX + "stylesrc.unsafe.desc"),
+                                        MESSAGE_PREFIX + "stylesrc.unsafe.otherinfo"),
                                 "",
                                 Alert.RISK_MEDIUM,
-                                csp);
+                                csp,
+                                "6");
                     }
                 }
             }
@@ -418,18 +424,25 @@ public class ContentSecurityPolicyScanRule extends PluginPassiveScanner {
     }
 
     private void raiseAlert(
-            String name, String description, String param, int risk, String evidence) {
+            String name,
+            String otherInfo,
+            String param,
+            int risk,
+            String evidence,
+            String alertRef) {
         String alertName = StringUtils.isEmpty(name) ? getName() : getName() + ": " + name;
 
         newAlert()
                 .setName(alertName)
                 .setRisk(risk)
                 .setConfidence(Alert.CONFIDENCE_MEDIUM)
-                .setDescription(description)
+                .setDescription(Constant.messages.getString(MESSAGE_PREFIX + "desc"))
+                .setOtherInfo(otherInfo)
                 .setParam(param)
                 .setSolution(getSolution())
                 .setReference(getReference())
                 .setEvidence(evidence)
+                .setAlertRef(PLUGIN_ID + "-" + alertRef)
                 .setCweId(getCweId())
                 .setWascId(getWascId())
                 .raise();