+++ addOns/pscanrules/src/test/java/org/zaproxy/zap/extension/pscanrules/ContentSecurityPolicyScanRuleUnitTest.java
@@ -20,12 +20,15 @@
 package org.zaproxy.zap.extension.pscanrules;
 
 import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.containsString;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.is;
+import static org.hamcrest.Matchers.not;
 
 import java.util.Map;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
 import org.parosproxy.paros.core.scanner.Alert;
 import org.parosproxy.paros.core.scanner.Plugin.AlertThreshold;
@@ -108,7 +111,7 @@ class ContentSecurityPolicyScanRuleUnitTest
 
         assertThat(alertsRaised.get(0).getName(), equalTo("CSP: Notices"));
         assertThat(
-                alertsRaised.get(0).getDescription(),
+                alertsRaised.get(0).getOtherInfo(),
                 equalTo(
                         "Warnings:\nUnrecognized directive default-src:\nUnrecognized directive report_uri\n"));
         assertThat(
@@ -119,7 +122,7 @@ class ContentSecurityPolicyScanRuleUnitTest
 
         assertThat(alertsRaised.get(1).getName(), equalTo("CSP: Wildcard Directive"));
         assertThat(
-                alertsRaised.get(1).getDescription(),
+                alertsRaised.get(1).getOtherInfo(),
                 equalTo(
                         "The following directives either allow wildcard sources (or ancestors), are not "
                                 + "defined, or are overly broadly defined: \nscript-src, style-src, img-src, "
@@ -132,6 +135,27 @@ class ContentSecurityPolicyScanRuleUnitTest
                 equalTo("default-src: 'none'; report_uri /__cspreport__"));
         assertThat(alertsRaised.get(1).getRisk(), equalTo(Alert.RISK_MEDIUM));
         assertThat(alertsRaised.get(1).getConfidence(), equalTo(Alert.CONFIDENCE_MEDIUM));
+        assertThat(alertsRaised.get(1).getAlertRef(), equalTo("10055-4"));
+    }
+
+    @Test
+    void shouldNotAlertOnValidSyntaxWhenCspContainsSyntaxIssues() {
+        // Given
+        HttpMessage msg =
+                createHttpMessage(
+                        "default-src: 'none'; report_uri /__cspreport__; frame-ancestors 'none'");
+        // When
+        scanHttpResponseReceive(msg);
+        // Then
+        assertThat(alertsRaised.size(), equalTo(2));
+        assertThat(alertsRaised.get(1).getName(), equalTo("CSP: Wildcard Directive"));
+        assertThat(alertsRaised.get(1).getOtherInfo(), not(containsString("frame-ancestors")));
+        assertThat(
+                alertsRaised.get(1).getEvidence(),
+                equalTo("default-src: 'none'; report_uri /__cspreport__; frame-ancestors 'none'"));
+        assertThat(alertsRaised.get(1).getRisk(), equalTo(Alert.RISK_MEDIUM));
+        assertThat(alertsRaised.get(1).getConfidence(), equalTo(Alert.CONFIDENCE_MEDIUM));
+        assertThat(alertsRaised.get(1).getAlertRef(), equalTo("10055-4"));
     }
 
     @Test
@@ -145,7 +169,7 @@ class ContentSecurityPolicyScanRuleUnitTest
 
         assertThat(alertsRaised.get(0).getName(), equalTo("CSP: Notices"));
         assertThat(
-                alertsRaised.get(0).getDescription(),
+                alertsRaised.get(0).getOtherInfo(),
                 equalTo(
                         "Warnings:\nThis host name is unusual, and likely meant to be a keyword that is missing the required quotes: 'none'.\n"));
 
@@ -154,6 +178,7 @@ class ContentSecurityPolicyScanRuleUnitTest
                 equalTo("default-src none; report-to csp-endpoint"));
         assertThat(alertsRaised.get(0).getRisk(), equalTo(Alert.RISK_LOW));
         assertThat(alertsRaised.get(0).getConfidence(), equalTo(Alert.CONFIDENCE_MEDIUM));
+        assertThat(alertsRaised.get(0).getAlertRef(), equalTo("10055-3"));
     }
 
     @Test
@@ -168,7 +193,7 @@ class ContentSecurityPolicyScanRuleUnitTest
 
         assertThat(alertsRaised.get(0).getName(), equalTo("CSP: Wildcard Directive"));
         assertThat(
-                alertsRaised.get(0).getDescription(),
+                alertsRaised.get(0).getOtherInfo(),
                 equalTo(
                         "The following directives either allow wildcard sources (or ancestors), are not "
                                 + "defined, or are overly broadly defined: \nframe-ancestors"
@@ -179,6 +204,7 @@ class ContentSecurityPolicyScanRuleUnitTest
                 equalTo("frame-ancestors *; default-src 'self'; form-action 'none'"));
         assertThat(alertsRaised.get(0).getRisk(), equalTo(Alert.RISK_MEDIUM));
         assertThat(alertsRaised.get(0).getConfidence(), equalTo(Alert.CONFIDENCE_MEDIUM));
+        assertThat(alertsRaised.get(0).getAlertRef(), equalTo("10055-4"));
     }
 
     @Test
@@ -213,17 +239,18 @@ class ContentSecurityPolicyScanRuleUnitTest
         assertThat(alertsRaised.size(), equalTo(2));
         assertThat(alertsRaised.get(0).getName(), equalTo("CSP: Notices"));
         assertThat(
-                alertsRaised.get(0).getDescription(),
+                alertsRaised.get(0).getOtherInfo(),
                 equalTo(
                         "Warnings:\nThis host name is unusual, and likely meant to be a keyword that is missing the required quotes: 'none'.\n"));
         assertThat(alertsRaised.get(0).getEvidence(), equalTo(policy));
         assertThat(alertsRaised.get(0).getRisk(), equalTo(Alert.RISK_LOW));
         assertThat(alertsRaised.get(0).getConfidence(), equalTo(Alert.CONFIDENCE_MEDIUM));
+        assertThat(alertsRaised.get(0).getAlertRef(), equalTo("10055-3"));
     }
 
     @ParameterizedTest
-    @ValueSource(strings = {"X-Content-Security-Policy", "X-WebKit-CSP"})
-    void shouldRaiseAlertOnLegacyCspHeader(String input) {
+    @CsvSource(value = {"X-Content-Security-Policy, 1", "X-WebKit-CSP, 2"})
+    void shouldRaiseAlertOnLegacyCspHeader(String input, String alertRef) {
         // Given
         HttpMessage msg = createHttpMessageWithReasonableCsp(input);
         // When
@@ -233,6 +260,7 @@ class ContentSecurityPolicyScanRuleUnitTest
         assertThat(alertsRaised.get(0).getName(), equalTo("CSP: " + input));
         assertThat(alertsRaised.get(0).getRisk(), equalTo(Alert.RISK_LOW));
         assertThat(alertsRaised.get(0).getConfidence(), equalTo(Alert.CONFIDENCE_MEDIUM));
+        assertThat(alertsRaised.get(0).getAlertRef(), equalTo("10055-" + alertRef));
     }
 
     @Test
@@ -247,6 +275,7 @@ class ContentSecurityPolicyScanRuleUnitTest
         assertThat(alertsRaised.get(1).getName(), equalTo("CSP: script-src unsafe-inline"));
         assertThat(alertsRaised.get(1).getRisk(), equalTo(Alert.RISK_MEDIUM));
         assertThat(alertsRaised.get(1).getConfidence(), equalTo(Alert.CONFIDENCE_MEDIUM));
+        assertThat(alertsRaised.get(1).getAlertRef(), equalTo("10055-5"));
     }
 
     @Test
@@ -261,6 +290,7 @@ class ContentSecurityPolicyScanRuleUnitTest
         assertThat(alertsRaised.get(1).getName(), equalTo("CSP: style-src unsafe-inline"));
         assertThat(alertsRaised.get(1).getRisk(), equalTo(Alert.RISK_MEDIUM));
         assertThat(alertsRaised.get(1).getConfidence(), equalTo(Alert.CONFIDENCE_MEDIUM));
+        assertThat(alertsRaised.get(1).getAlertRef(), equalTo("10055-6"));
     }
 
     @Test
@@ -275,11 +305,13 @@ class ContentSecurityPolicyScanRuleUnitTest
         // Verify the specific alerts
         assertThat(alertsRaised.get(0).getName(), equalTo("CSP: Wildcard Directive"));
         assertThat(alertsRaised.get(0).getEvidence(), equalTo("style-src 'unsafe-inline'"));
+        assertThat(alertsRaised.get(0).getAlertRef(), equalTo("10055-4"));
 
         assertThat(alertsRaised.get(1).getName(), equalTo("CSP: style-src unsafe-inline"));
         assertThat(alertsRaised.get(1).getRisk(), equalTo(Alert.RISK_MEDIUM));
         assertThat(alertsRaised.get(1).getConfidence(), equalTo(Alert.CONFIDENCE_MEDIUM));
         assertThat(alertsRaised.get(1).getEvidence(), equalTo("style-src 'unsafe-inline'"));
+        assertThat(alertsRaised.get(1).getAlertRef(), equalTo("10055-6"));
     }
 
     private HttpMessage createHttpMessageWithReasonableCsp(String cspHeaderName) {